move services and stores to domain layer
