+++ src/domain/services/user-services.ts
@@ -1,6 +1,6 @@
 import userModel from "../stores/user-store";
-import {createToken} from "../utils/utils";
-import {usernameExists, emailExists} from "../utils/utils";
+import {createToken} from "../../utils/utils";
+import {usernameExists, emailExists} from "../../utils/utils";
 import * as bcrypt from 'bcrypt'
