+++ src/domain/stores/user-store.ts
@@ -1,7 +1,7 @@
 import { PrismaClient} from "@prisma/client";
 import { v4 as uuidv4 } from "uuid";
 import * as bcrypt from "bcrypt";
-import { createToken } from "../utils/utils";
+import { createToken } from "../../utils/utils";
 
 const prisma = new PrismaClient();
